Normalize inverse results to Int at the InverseMod boundary
The contract documented on this branch said an InverseMod may return any
Integer. The rest of the package is Int arithmetic throughout, so that was
not true, and the failure was silent: measure! forms

    a = mod(-commutator * tab.inversemod(comm0, d), d)

and with an unsigned inverse the negation and multiply evaluate in unsigned
arithmetic and wrap BEFORE the mod. At d = 3 with PrecomputedInvMod(UInt64)
both tableau types returned normally and were left holding two pairs of
non-commuting generators -- a corrupt state, no error. An Int32 table instead
threw a MethodError from binom2_mod_oddprime(::Int32, ::Int64, ::Int32).

Both predate this branch; what the branch added was a contract and a test
asserting unsigned tables are legal, turning a latent bug into a documented
guarantee.

Fixed by converting at the call operator with a ::Int return annotation,
rather than by routing every consumer through a helper. The annotation is the
one point every consumer already passes through, so there are no call sites
to audit and a new one cannot bypass it. For an Int table the conversion is
the identity.

Integer-backed storage of any width stays accepted -- Int32 and UInt64 tables
now agree exactly with an Int control on canonicalization, expectation, span
reconstruction, measurement, duality and the x.z cache. The comments claiming
the table's element type deliberately widens downstream arithmetic are gone;
that was never compatible with the package's Int phase helpers.

+++ src/inversemod.jl
@@ -1,20 +1,26 @@
 ##### modular-inverse strategies #####
 #
 # CONTRACT: an `InverseMod` maps `(x, d)` with `0 < x < d` to the inverse of `x`
-# modulo `d`, and **must return an `Integer`**.
+# modulo `d`, and **every consumer receives an `Int`**.
 #
-# The returned value's type is not incidental. It is the type the downstream
-# multiplications are evaluated in -- `src/modular.jl` forms `a * src[i]` with
-# `a` exactly as returned -- so the table's element type also sets the overflow
-# ceiling: an `Int` table caps the package where `(d-1)^2` passes `typemax(Int)`,
-# while an unsigned table of the same width raises it.
+# The lookup table may be stored in any integer type -- `Vector{Int32}` and
+# `Vector{UInt64}` are both accepted, and a lazy `AbstractVector` works too --
+# but the value is converted on read, at the call operator below, so no caller
+# ever has to think about it. A valid inverse lies in `1:(d-1)` and `d` is an
+# `Int`, so the conversion is always exact.
+#
+# That normalization is not cosmetic. The rest of the package is `Int`
+# arithmetic throughout: `binom2_mod_oddprime` takes `Int`, and `measure!`
+# forms `mod(-commutator * inv, d)`. Handed an unsigned inverse, that negation
+# and multiply evaluate in unsigned arithmetic and wrap BEFORE the `mod`,
+# which silently leaves the tableau with non-commuting generators -- no error,
+# just a corrupt state. Converting once, here, is what prevents that.
 #
 # Non-integer tables are rejected at construction. An exactly-valued `Float64`
 # table happens to work, because the result converts on assignment into the
 # `Int` tableau, but one off by a single ULP would round into a wrong-but-valid
-# residue -- a silently incorrect state, which is far worse than a construction
-# error. The check is at the constructor so the failure lands where the mistake
-# is, rather than as a MethodError inside an inner loop.
+# residue. The check is at the constructor so the failure lands where the
+# mistake is.
 
 abstract type InverseMod end
 
@@ -36,5 +42,8 @@ struct PrecomputedInvMod{T <: AbstractVector} <: InverseMod
     end
 end
 
-@inline (jit::JustInTimeInvMod)(x::Int, d::Int) = Base.invmod(x, d)
-@inline (pre::PrecomputedInvMod)(x::Int, d::Int) = (@inbounds pre.lookuptable[x])
+# The `::Int` return annotations are the normalization point for the whole
+# package; see the contract note above. For an `Int` table the conversion is
+# the identity, so the hot path is unchanged.
+@inline (jit::JustInTimeInvMod)(x::Int, d::Int)::Int = Base.invmod(x, d)
+@inline (pre::PrecomputedInvMod)(x::Int, d::Int)::Int = (@inbounds pre.lookuptable[x])
